Move AppComponentContext to feature subproject
Remove architectural dependency on AppComponentContext. The `arkitekt-decompose` should be unaware of application-specific context.
