+++ shared/arkitekt-decompose/src/commonMain/kotlin/app/futured/arkitekt/decompose/navigation/Destination.kt
@@ -1,10 +1,10 @@
 package app.futured.arkitekt.decompose.navigation
 
-import app.futured.arkitekt.decompose.AppComponentContext
+import com.arkivanov.decompose.GenericComponentContext
 
 /**
  * A base interface for destinations used in Decompose navigation.
  */
-interface Destination<out C : NavEntry> {
-    fun createComponent(componentContext: AppComponentContext): C
+interface Destination<out C : NavEntry, CTX: GenericComponentContext<CTX>> {
+    fun createComponent(componentContext: CTX): C
 }
