+++ shared/feature/src/commonMain/kotlin/app/futured/kmptemplate/feature/navigation/root/RootNavigationFactory.kt
@@ -1,6 +1,6 @@
 package app.futured.kmptemplate.feature.navigation.root
 
-import app.futured.arkitekt.decompose.AppComponentContext
+import app.futured.kmptemplate.feature.AppComponentContext
 
 object RootNavigationFactory {
     fun create(componentContext: AppComponentContext): RootNavigation = RootNavigationComponent(componentContext)
