+++ shared/feature/src/commonMain/kotlin/app/futured/kmptemplate/feature/navigation/signedin/tab/a/TabANavigationComponent.kt
@@ -1,5 +1,5 @@
 package app.futured.kmptemplate.feature.navigation.signedin.tab.a
 
-import app.futured.arkitekt.decompose.AppComponentContext
+import app.futured.kmptemplate.feature.AppComponentContext
 
 internal class TabANavigationComponent(componentContext: AppComponentContext) : TabANavigation, AppComponentContext by componentContext
