+++ shared/feature/src/commonMain/kotlin/app/futured/kmptemplate/feature/navigation/signedin/tab/b/TabBNavigator.kt
@@ -1,8 +1,8 @@
 package app.futured.kmptemplate.feature.navigation.signedin.tab.b
 
-import app.futured.arkitekt.decompose.AppComponentContext
 import app.futured.arkitekt.decompose.ext.asStateFlow
 import app.futured.arkitekt.decompose.ext.componentCoroutineScope
+import app.futured.kmptemplate.feature.AppComponentContext
 import app.futured.kmptemplate.feature.navigation.root.RootNavigator
 import com.arkivanov.decompose.Child
 import com.arkivanov.decompose.router.stack.ChildStack
