+++ shared/feature/src/commonMain/kotlin/app/futured/kmptemplate/feature/navigation/signedin/tab/c/TabCNavigationComponent.kt
@@ -1,5 +1,5 @@
 package app.futured.kmptemplate.feature.navigation.signedin.tab.c
 
-import app.futured.arkitekt.decompose.AppComponentContext
+import app.futured.kmptemplate.feature.AppComponentContext
 
 internal class TabCNavigationComponent(componentContext: AppComponentContext) : TabCNavigation, AppComponentContext by componentContext
