+++ shared/feature/src/commonMain/kotlin/app/futured/kmptemplate/feature/ui/first/FirstComponent.kt
@@ -1,7 +1,7 @@
 package app.futured.kmptemplate.feature.ui.first
 
-import app.futured.arkitekt.decompose.AppComponentContext
-import app.futured.arkitekt.decompose.viewModel
+import app.futured.arkitekt.decompose.injection.viewModel
+import app.futured.kmptemplate.feature.AppComponentContext
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.StateFlow
