+++ shared/feature/src/commonMain/kotlin/app/futured/kmptemplate/feature/ui/login/LoginComponent.kt
@@ -1,7 +1,7 @@
 package app.futured.kmptemplate.feature.ui.login
 
-import app.futured.arkitekt.decompose.AppComponentContext
-import app.futured.arkitekt.decompose.viewModel
+import app.futured.arkitekt.decompose.injection.viewModel
+import app.futured.kmptemplate.feature.AppComponentContext
 import kotlinx.coroutines.flow.StateFlow
 
 internal class LoginComponent(
